requirements.txt
Flask==2.3.3
pytesseract==0.3.10
Pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
Werkzeug==2.3.7
PyMuPDF==1.23.8 